user,role controller revised with authentication and authorization

+++ src/main/java/com/nirmal/personalfinancetracker/Controller/UserController.java
@@ -1,5 +1,6 @@
 package com.nirmal.personalfinancetracker.Controller;
 
+import com.nirmal.personalfinancetracker.dto.request.UserRequestDto;
 import com.nirmal.personalfinancetracker.dto.response.Response;
 import com.nirmal.personalfinancetracker.dto.response.UserResponseDto;
 import com.nirmal.personalfinancetracker.model.User;
@@ -8,6 +9,9 @@ import com.nirmal.personalfinancetracker.service.impl.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,25 +23,16 @@ public class UserController {
     private UserServiceImpl userServiceImpl;
     @Autowired
     private DtoMapperImpl dtoMapper;
-
-    @PostMapping("/user")
-    public ResponseEntity<Response<UserResponseDto>> registerUser(@RequestBody User user) {
-        Response<UserResponseDto> response = new Response<>();
-        UserResponseDto userResponseDto = userServiceImpl.registerUser(user);
-        response.setStatus(true);
-        response.setData(userResponseDto);
-        response.setMessage("user registered successfully");
-        return new ResponseEntity<>(response, HttpStatus.OK);
-    }
+    @Autowired
+    private PasswordEncoder passwordEncoder;
 
     @GetMapping("/users")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Response<List<UserResponseDto>>> viewUserList() {
         Response<List<UserResponseDto>> response = new Response<>();
         List<UserResponseDto> userResponseDtos = userServiceImpl.viewUsers();
         if (!userResponseDtos.isEmpty()) {
-            response.setData(userResponseDtos);
-            response.setStatus(true);
-            response.setMessage("list retrieved successfully");
+            response.successResponse(userResponseDtos, "list retrieved successfully");
             return new ResponseEntity<>(response, HttpStatus.OK);
         }
         response.setStatus(false);
@@ -47,6 +42,7 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Response<UserResponseDto>> viewUserById(@PathVariable int id) {
         Response<UserResponseDto> response = new Response<>();
         UserResponseDto user = userServiceImpl.viewUserById(id);
@@ -63,26 +59,34 @@ public class UserController {
     }
 
     @PutMapping("/user/{id}")
-    public ResponseEntity<Response<UserResponseDto>> editUser(@PathVariable int id, @RequestBody User user) {
+    @PreAuthorize("hasAnyRole('ADMIN','USER')")
+    public ResponseEntity<Response<UserResponseDto>> editUser(@PathVariable int id, @RequestBody UserRequestDto userRequestDto, Authentication authentication) {
         Response<UserResponseDto> response = new Response<>();
-        UserResponseDto user1 = userServiceImpl.updateUser(id, user);
-        if (user1 != null) {
-            response.successResponse(user1,"user updated successfully");
-            return new ResponseEntity<>(response, HttpStatus.OK);
+        if(id == ((User)authentication.getPrincipal()).getId()) {
+            User user = dtoMapper.toUserEntity(userRequestDto);
+            user.setPassword(passwordEncoder.encode(userRequestDto.getPassword()));
+            UserResponseDto user1 = userServiceImpl.updateUser(id, user);
+            if (user1 != null) {
+                response.successResponse(user1, "user updated successfully");
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            }
         }
-        response.failureResponse("Id invalid");
-        return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        response.failureResponse("User not Authorized");
+        return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/user/{id}")
-    public ResponseEntity<Response<String>> deleteUser(@PathVariable int id) {
+    @PreAuthorize("hasAnyRole('ADMIN','USER')")
+    public ResponseEntity<Response<String>> deleteUser(@PathVariable int id, Authentication authentication) {
         Response<String> response = new Response<>();
-        String data = userServiceImpl.deleteUser(id);
-        if (data.equals("success")) {
-            response.successResponse(data,"user deleted successfully");
-            return new ResponseEntity<>(response, HttpStatus.OK);
+        if(id == ((User)authentication.getPrincipal()).getId()) {
+            String data = userServiceImpl.deleteUser(id);
+            if (data.equals("success")) {
+                response.successResponse(data, "user deleted successfully");
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            }
         }
-        response.failureResponse(data);
-        return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        response.failureResponse("User not Authorized");
+        return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
     }
 }
